Rearranging functions/files.
