+++ surveySimPP/__init__.py
@@ -1,2 +1,3 @@
 from . import modules
 from . import lsstcomet
+from . import surveySimPP_imports
