+++ surveySimPP/modules/__init__.py
@@ -24,7 +24,10 @@ from . import PPReadOif
 from . import PPReadOrbitFile
 from . import PPTrailingLoss
 from . import PPVignetting
-from . import PPRunUtilities
+from . import PPConfigParser
+from . import PPCMDLineParser
+from . import PPReadAllInput
+from . import PPGetLogger
 from . import PPApplyFOVFilter
 from . import PPSNRLimit
 from . import PPFilterFadingFunction
