+++ surveySimPP/modules/tests/test_PPConfigParser.py
@@ -28,7 +28,7 @@ def setup_and_teardown_for_PPConfigFileParser(tmp_path):
 
 def test_PPConfigFileParser(setup_and_teardown_for_PPConfigFileParser):
 
-    from surveySimPP.modules.PPRunUtilities import PPConfigFileParser
+    from surveySimPP.modules.PPConfigParser import PPConfigFileParser
 
     configs = PPConfigFileParser(get_test_filepath('test_PPConfig.ini'), 'lsst')
