+++ surveySimPP/surveySimPP.py
@@ -4,20 +4,8 @@ import sys
 import time
 import numpy as np
 import argparse
-from surveySimPP.modules.PPMatchPointing import PPMatchPointing
-from surveySimPP.modules.PPFilterSSPLinking import PPFilterSSPLinking
-from surveySimPP.modules.PPTrailingLoss import PPTrailingLoss
-from surveySimPP.modules.PPBrightLimit import PPBrightLimit
-from surveySimPP.modules.PPMakeIntermediatePointingDatabase import PPMakeIntermediatePointingDatabase
-from surveySimPP.modules.PPCalculateApparentMagnitude import PPCalculateApparentMagnitude
-from surveySimPP.modules.PPApplyFOVFilter import PPApplyFOVFilter
-from surveySimPP.modules.PPSNRLimit import PPSNRLimit
-from surveySimPP.modules import PPAddUncertainties, PPRandomizeMeasurements
-from surveySimPP.modules import PPVignetting
-from surveySimPP.modules.PPFilterFadingFunction import PPFilterFadingFunction
-from surveySimPP.modules.PPRunUtilities import PPGetLogger, PPConfigFileParser, PPPrintConfigsToLog, PPCMDLineParser, PPReadAllInput
-from surveySimPP.modules.PPMagnitudeLimit import PPMagnitudeLimit
-from surveySimPP.modules.PPOutput import PPWriteOutput
+
+from surveySimPP.surveySimPP_imports import *
 
 
 # Author: Samuel Cornwall, Siegfried Eggl, Grigori Fedorets, Steph Merritt, Meg Schwamb
